+++ src/main/kotlin/com/supa/spring/supaspring/file/domain/SupabaseFileService.kt
@@ -1,0 +1,25 @@
+package com.supa.spring.supaspring.file.domain
+
+import com.supa.spring.supaspring.file.application.BucketItemDto
+import com.supa.spring.supaspring.file.infrastructure.SupabaseFileRepository
+import org.springframework.stereotype.Service
+import org.springframework.web.multipart.MultipartFile
+
+@Service
+class SupabaseFileService(val supabaseFileRepository: SupabaseFileRepository) : FileService {
+    override fun getFilesInBucket(bucketId: String): List<BucketItemDto> {
+        return supabaseFileRepository.getFilesInBucket(bucketId)
+    }
+
+    override fun moveFile(from: String, to: String) {
+
+    }
+
+    override fun uploadFile(bucketId: String, file: MultipartFile) {
+
+    }
+
+    override fun replaceFileAtPath(path: String, file: MultipartFile) {
+
+    }
+}
